+++ core/processes.php
@@ -2,7 +2,7 @@
 
 <?php if ($_SESSION['usertype'] == 'Administrators' || $_SESSION['usertype'] == 'Operators' || $_SESSION['usertype'] == 'Users' || $_SESSION['usersett']['processcontrol'] > 0) {  } else { header('location: /'); exit; } ?>
 
-<?php if (!isset($_GET['csv_processes'])) { $_SESSION['processes'] = $_SERVER['REQUEST_URI']; } ?>
+<?php if (!isset($_GET['csv_processes'])) { $_SESSION['processes'] = htmlspecialchars((string)$_SERVER['REQUEST_URI'], ENT_QUOTES, 'UTF-8'); } ?>
 
 <?php include("navigation.php"); ?>
 
@@ -30,7 +30,7 @@ foreach($wmicpu as $cpu) {
 	function endprocess(IDProcess,Name,PercentProcessorTime,WorkingSetPrivate,CreatingProcessID,LimitName){
 		$.ajax({
 			type: "GET",
-			url: 'processesjqsrv.php?idprocess=' + IDProcess + '&pid=' + CreatingProcessID,
+			url: 'processesjqsrv.php?<?php echo substr(md5('$_GET' . $sessiontoken), 5, 15) . '=' . substr(md5('$_GET' . $sessiontoken), 15, 25); ?>&idprocess=' + IDProcess + '&pid=' + CreatingProcessID,
 			data: '',
 			dataType: 'json',
 			cache: false,
@@ -145,7 +145,7 @@ foreach($wmicpu as $cpu) {
 
 					$message = '';
 					
-					if ($endidprocess != '' && $endtypeprocess != '' && (strpos($_SERVER['HTTP_REFERER'], $_SERVER['HTTP_HOST'] . $_SERVER['SCRIPT_NAME'])) > 0) {
+					if ($endidprocess != '' && $endtypeprocess != '' && strpos($_SERVER['HTTP_REFERER'], $_SERVER['HTTP_HOST'] . $_SERVER['SCRIPT_NAME']) > 0 && isset($_POST[substr(md5('$_POST' . $_SESSION['tokenl']), 5, 15)])) {
 						if ($endtypeprocess == 1) {
 							session_write_close();
 							$endprocessoutput = exec('taskkill.exe /f /pid ' . $endidprocess, $errorarray, $errorlevel);
@@ -173,7 +173,11 @@ foreach($wmicpu as $cpu) {
 					}
 
 					if (isset($_GET['page'])) {
-						$pgkey = $_GET['page'] - 1;
+						if (is_numeric($_GET['page'])) {
+							$pgkey = $_GET['page'] - 1;
+						} else {
+							$pgkey = 0;
+						}
 					} else {
 						$pgkey = 0;
 					}
@@ -185,14 +189,15 @@ foreach($wmicpu as $cpu) {
                     	<tr><td width="20%"><div id="csvexport"></div></td><td width="80%"><div id="totalelement" style="font-size:12px;"></div></td></tr>
                     	<tr><td width="20%"><div style="font-size:12px; white-space:nowrap; table-layout:fixed; overflow:hidden;">Page Loading Time:</div></td><td width="80%"><div id="totaltime" style="font-size:12px;"></div></td></tr>
                     	<tr><td width="20%"><div style="font-size:12px; white-space:nowrap; table-layout:fixed; overflow:hidden;">Reloading Time:</div></td><td width="80%" style="font-size:12px;"><?php if ($processesrrsetting != 'Hold') { echo number_format(($processesrrsetting / 1000), 0, ',', '.') . '&nbsp;sec&nbsp;&nbsp;'; } else { echo $processesrrsetting . '&nbsp;&nbsp;'; } ?><a href="?orderby=<?php echo $orderby; ?>&cpucount=<?php echo $cpucount; ?>&filter=<?php echo urlencode($filter); ?>&page=<?php echo $pgkey + 1; ?>" title="Reload Now"><div class="icon-loop"></div></a></td></tr>
-						<?php if ($filter != '') { ?><tr><td width="20%"><div style="font-size:12px; white-space:nowrap; table-layout:fixed; overflow:hidden;">Filter:</div></td><td width="80%" style="font-size:12px;"><i><?php echo $filter; ?></i></td></tr><?php } ?>
+						<?php if ($filter != '') { ?><tr><td width="20%"><div style="font-size:12px; white-space:nowrap; table-layout:fixed; overflow:hidden;">Filter:</div></td><td width="80%" style="font-size:12px;"><i><?php echo str_replace(')', '&rpar;', str_replace('(', '&lpar;', str_replace('=', '&equals;', htmlspecialchars((string)$filter, ENT_QUOTES, 'UTF-8')))); ?></i></td></tr><?php } ?>
                     </table>
                     
 					<div style="font-size:12px; white-space:nowrap; table-layout:fixed; overflow:hidden;">
                     <blockquote style="font-size:12px; height:33px;" title="<?php echo 'Use Normal String for SIMPLE SEARCH' . "\n" . 'Use Regular Expression for COMPLEX SEARCH' . "\n" . 'Use Minus  -  for NOT CONTAIN' . "\n" . 'Use Pipe  |  for OR OPERATOR' . "\n" . 'Use Raw Data View for REFERENCES'; ?>">
                     	<form id="filterform" name="filterform" method="get">
-                        	Filter:&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;<input type="text" id="filter" name="filter" placeholder="Regular Expression..." value="<?php echo $filter; ?>" title="<?php echo $filter; ?>" style="font-family:\'Segoe UI Light\',\'Open Sans\',Verdana,Arial,Helvetica,sans-serif; font-size:12px; border:solid; border-width:1px; border-color:#e5e5e5; width:170px; height:23px; padding-top:0px; padding-left:4px; padding-right:4px;" />&nbsp;&nbsp;<a href="javascript:;" onClick="document.getElementById('filterform').submit();" title="Filter by String or Regular Expression"><div class="icon-search"<?php if ($filter != '') { echo ' style="color:#8063C8;"'; } ?>></div></a><?php if ($filter != '') { ?>&nbsp;<a href="?orderby=<?php echo $orderby; ?>" title="Clear Filter"><div class="icon-cancel"></div></a><?php } ?>
+                        	Filter:&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;<input type="text" id="filter" name="filter" placeholder="Regular Expression..." value="<?php echo str_replace(')', '&rpar;', str_replace('(', '&lpar;', str_replace('=', '&equals;', htmlspecialchars((string)$filter, ENT_QUOTES, 'UTF-8')))); ?>" title="<?php echo str_replace(')', '&rpar;', str_replace('(', '&lpar;', str_replace('=', '&equals;', htmlspecialchars((string)$filter, ENT_QUOTES, 'UTF-8')))); ?>" style="font-family:\'Segoe UI Light\',\'Open Sans\',Verdana,Arial,Helvetica,sans-serif; font-size:12px; border:solid; border-width:1px; border-color:#e5e5e5; width:170px; height:23px; padding-top:0px; padding-left:4px; padding-right:4px;" />&nbsp;&nbsp;<a href="javascript:;" onClick="document.getElementById('filterform').submit();" title="Filter by String or Regular Expression"><div class="icon-search"<?php if ($filter != '') { echo ' style="color:#8063C8;"'; } ?>></div></a><?php if ($filter != '') { ?>&nbsp;<a href="?orderby=<?php echo $orderby; ?>" title="Clear Filter"><div class="icon-cancel"></div></a><?php } ?>
                             <input type="hidden" id="orderby" name="orderby" value="<?php echo $orderby; ?>" />
+							<input type="hidden" id="<?php echo substr(md5('$_GET' . $sessiontoken), 5, 15); ?>" name="<?php echo substr(md5('$_GET' . $sessiontoken), 5, 15); ?>" value="<?php echo substr(md5('$_GET' . $sessiontoken), 15, 25); ?>" />
 						</form>
 					</blockquote>
 					</div>
@@ -209,7 +214,7 @@ foreach($wmicpu as $cpu) {
 					function update() {
 						$.ajax({
 							type: "GET",
-							url: 'processesjq.php?orderby=<?php echo $orderby; ?>&cpucount=<?php echo $cpucount; ?>&filter=<?php echo urlencode($filter); ?>&page=<?php echo $pgkey; ?>&phptimeout=<?php echo $phptimeout; ?>',
+							url: 'processesjq.php?<?php echo substr(md5('$_GET' . $sessiontoken), 5, 15) . '=' . substr(md5('$_GET' . $sessiontoken), 15, 25); ?>&orderby=<?php echo $orderby; ?>&cpucount=<?php echo $cpucount; ?>&filter=<?php echo urlencode($filter); ?>&page=<?php echo $pgkey; ?>&phptimeout=<?php echo $phptimeout; ?>',
 							data: '',
 							dataType: 'json',
 							cache: false,
@@ -229,6 +234,7 @@ foreach($wmicpu as $cpu) {
 						<input type="hidden" id="endidprocess" name="endidprocess" />
 						<input type="hidden" id="endnameprocess" name="endnameprocess" />
 						<input type="hidden" id="endtypeprocess" name="endtypeprocess" />
+						<input type="hidden" id="<?php echo substr(md5('$_POST' . $sessiontoken), 5, 15); ?>" name="<?php echo substr(md5('$_POST' . $sessiontoken), 5, 15); ?>" value="<?php echo substr(md5('$_POST' . $sessiontoken), 15, 25); ?>" />
 					</form>
 
 					</div>
